requirements.txt
# Web Framework
Flask==3.0.3
Flask-SocketIO==5.3.7

# Audio Processing
PyAudio==0.2.14
pydub==0.25.1
silero-vad==5.1
sounddevice==0.5.0
soundfile==0.12.1
playsound==1.3.0
pygame==2.6.1

# Speech Recognition and Synthesis
funasr==1.1.6
edge-tts==6.1.12
gTTS==2.5.3
chattts==0.1.1

# AI and Machine Learning
torch==2.4.1
torchaudio==2.4.1
openai==1.45.0
langchain==0.3.2
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0

# HTTP and Network
requests==2.32.3
certifi>=2024.2.2
urllib3>=2.2.0
beautifulsoup4==4.12.2

# Utilities
PyYAML==6.0.2
schedule==1.2.2
numpy>=1.26.4

# Optional but recommended for better performance
# uncomment if needed
# cuda-python>=12.3.0  # for NVIDIA GPU support
